remove output_resources
