+++ pkg/reconciler/pipelinerun/resources/input_output_steps.go
@@ -17,41 +17,9 @@ limitations under the License.
 package resources
 
 import (
-	"path/filepath"
-
 	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1beta1"
 	resourcev1alpha1 "github.com/tektoncd/pipeline/pkg/apis/resource/v1alpha1"
 )
-
-// GetOutputSteps will add the correct `path` to the output resources for pt
-func GetOutputSteps(outputs map[string]*resourcev1alpha1.PipelineResource, taskName string) []v1beta1.TaskResourceBinding {
-	var taskOutputResources []v1beta1.TaskResourceBinding
-
-	for name, outputResource := range outputs {
-		taskOutputResource := v1beta1.TaskResourceBinding{
-			PipelineResourceBinding: v1beta1.PipelineResourceBinding{
-				Name: name,
-			},
-			Paths: []string{filepath.Join(taskName, name)},
-		}
-		// SelfLink is being checked there to determine if this PipelineResource is an instance that
-		// exists in the cluster (in which case Kubernetes will populate this field) or is specified by Spec
-		if outputResource.SelfLink != "" {
-			taskOutputResource.ResourceRef = &v1beta1.PipelineResourceRef{
-				Name:       outputResource.Name,
-				APIVersion: outputResource.APIVersion,
-			}
-		} else if outputResource.Spec.Type != "" {
-			taskOutputResource.ResourceSpec = &resourcev1alpha1.PipelineResourceSpec{
-				Type:         outputResource.Spec.Type,
-				Params:       outputResource.Spec.Params,
-				SecretParams: outputResource.Spec.SecretParams,
-			}
-		}
-		taskOutputResources = append(taskOutputResources, taskOutputResource)
-	}
-	return taskOutputResources
-}
 
 // GetInputSteps will add the correct `path` to the input resources for pt. If the resources are provided by
 // a previous task, the correct `path` will be used so that the resource provided by that task will be used.
@@ -96,7 +64,4 @@ func WrapSteps(tr *v1beta1.TaskRunSpec, pt *v1beta1.PipelineTask, inputs, output
 		// Add presteps to setup updated input
 		tr.Resources.Inputs = append(tr.Resources.Inputs, GetInputSteps(inputs, pt.Resources.Inputs)...)
 	}
-
-	// Add poststeps to setup outputs
-	tr.Resources.Outputs = append(tr.Resources.Outputs, GetOutputSteps(outputs, pt.Name)...)
 }
